Added Tests and made corrections to the models to reflect any errors.

+++ src/GetMatchDetails/Player.cs
@@ -8,7 +8,7 @@ namespace HGV.Daedalus.GetMatchDetails
 	public class Player
 	{
 		public int player_slot { get; set; }
-		public int account_id { get; set; }
+		public long account_id { get; set; }
 		public int hero_id { get; set; }
 		public int kills { get; set; }
 		public int death { get; set; }
